requirements.txt
paho-mqtt
pycryptodome>=3.20.0